+++ app/exception/unhandledRejection.js
@@ -1,0 +1,6 @@
+const unhandledRejection = async (err) => {
+  console.error(err)
+  process.exit(1)
+}
+
+module.exports = unhandledRejection
