+++ app/index.js
@@ -1,16 +1,11 @@
-const server = require('./server')
-const messageService = require('./messaging/service')
+const startServer = require('./server/init-server')
+const unhandledRejection = require('./exception/unhandledRejection')
+const { start: startMessageService, stop: stopMessageService } = require('./messaging/service')
 
-const init = async () => {
-  await server.start()
-  await messageService.start()
-  console.log('Server running on %s', server.info.uri)
-}
+startServer()
+startMessageService()
 
 process.on('unhandledRejection', async (err) => {
-  console.error(err)
-  await messageService.stop()
-  process.exit(1)
+  await stopMessageService()
+  unhandledRejection(err)
 })
-
-init()
