+++ app/register-your-interest/messaging/process-message.js
@@ -1,10 +1,6 @@
 const processRegisterYourInterestMessage = async (message) => {
-  try {
-    console.log(`Reading message from queue with body ${JSON.stringify(message.body)}`)
-    // todo https://dev.azure.com/defragovuk/DEFRA-FFC/_workitems/edit/135602
-  } catch (err) {
-    console.error('Unable to process register your interest request message:', err)
-  }
+  console.log(`Reading message from queue with body ${JSON.stringify(message.body)}`)
+  // todo https://dev.azure.com/defragovuk/DEFRA-FFC/_workitems/edit/135602
 }
 
 module.exports = processRegisterYourInterestMessage
