+++ app/server/init-server.js
@@ -1,0 +1,8 @@
+const server = require('./server')
+
+const startServer = async () => {
+  await server.start()
+  console.log('Server running on %s', server.info.uri)
+}
+
+module.exports = startServer
