+++ test/integration/narrow/app/server.test.js
@@ -1,6 +1,6 @@
 describe('Server test', () => {
   test('createServer returns server', () => {
-    const server = require('../../../../app/server')
+    const server = require('../../../../app/server/server')
     expect(server).toBeDefined()
   })
 })
