+++ test/integration/narrow/routes/healthz.test.js
@@ -1,5 +1,5 @@
 describe('Healthz test', () => {
-  const server = require('../../../../app/server')
+  const server = require('../../../../app/server/server')
 
   beforeEach(async () => {
     await server.start()
