--- test/unit/app/index.test.js
@@ -1,12 +1,0 @@
-describe('index test', () => {
-  test('Test server and message bus listener started', async () => {
-    // const server = require('../../../app/server')
-    // const messaging = require('../../../app/messaging/service')
-    // jest.mock('../../../app/server')
-    // jest.mock('../../../app/messaging/service')
-    // require('../../../app/index')
-
-    // // expect(messaging.start()).toHaveBeenCalled()
-    // // expect(server.start()).toHaveBeenCalled()
-  })
-})
