+++ test/unit/app/server/init-server.test.js
@@ -1,0 +1,9 @@
+describe('Init Server Test', () => {
+  test('Should start server', async () => {
+    const server = require('../../../../app/server/server')
+    jest.mock('../../../../app/server/server')
+    const startServer = require('../../../../app/server/init-server')
+    await startServer()
+    expect(server.start).toHaveBeenCalledTimes(1)
+  })
+})
